# C#: Implement object pools for Array, Dictionary, StringName, NodePath