--- a/modules/mono/glue/GodotSharp/GodotSharp/Core/Array.cs
+++ b/modules/mono/glue/GodotSharp/GodotSharp/Core/Array.cs
@@ -177,15 +177,20 @@ namespace Godot.Collections
 
         private Array(godot_array nativeValueToOwn)
         {
-            NativeValue = (godot_array.movable)(nativeValueToOwn.IsAllocated ?
-                nativeValueToOwn :
-                NativeFuncs.godotsharp_array_new());
-            _weakReferenceToSelf = DisposablesTracker.RegisterDisposable(this);
+            TakeOwnershipOfDisposableValue(nativeValueToOwn);
         }
 
         // Explicit name to make it very clear
         internal static Array CreateTakingOwnershipOfDisposableValue(godot_array nativeValueToOwn)
             => new Array(nativeValueToOwn);
+
+        internal void TakeOwnershipOfDisposableValue(godot_array nativeValueToOwn)
+        {
+            NativeValue = (godot_array.movable)(nativeValueToOwn.IsAllocated ?
+                       nativeValueToOwn :
+                       NativeFuncs.godotsharp_array_new());
+            _weakReferenceToSelf = DisposablesTracker.RegisterDisposable(this);
+        }
 
         ~Array()
         {
@@ -1087,7 +1092,7 @@ namespace Godot.Collections
             VariantUtils.GenericConversion<Array<T>>.FromVariantCb = FromVariantFunc;
         }
 
-        private readonly Array _underlyingArray;
+        internal Array _underlyingArray;
 
         Array IGenericGodotArray.UnderlyingArray => _underlyingArray;