Implement https://github.com/godotengine/godot-proposals/issues/12375

Benchmark the performance of `GetTree().GetNodesInGroup(cached_sn)` [cs-object-pools-alloc.zip](https://github.com/user-attachments/files/20061997/cs-object-pools-alloc.zip)
This PR:
```
// * Summary *

BenchmarkDotNet v0.14.0, EndeavourOS
AMD Ryzen 7 4800U with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.105
  [Host] : .NET 9.0.4 (9.0.425.16305), X64 RyuJIT AVX2

Toolchain=InGodotProcessNoEmitToolchain

| Method     | Mean      | Error    | StdDev   | Gen0      | Gen1      | Allocated |
|----------- |----------:|---------:|---------:|----------:|----------:|----------:|
| Base       | 140.20 ms | 2.563 ms | 6.040 ms | 1750.0000 | 1500.0000 |  11.44 MB |
| ReturnPool |  79.55 ms | 0.958 ms | 0.896 ms | 3000.0000 |         - |    6.1 MB |

```
Master:
```// * Summary *

BenchmarkDotNet v0.14.0, EndeavourOS
AMD Ryzen 7 4800U with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.105
  [Host] : .NET 9.0.4 (9.0.425.16305), X64 RyuJIT AVX2

Toolchain=InGodotProcessNoEmitToolchain  

| Method | Mean     | Error   | StdDev  | Gen0      | Gen1      | Allocated |
|------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Base   | 152.9 ms | 3.03 ms | 4.80 ms | 1750.0000 | 1500.0000 |  11.44 MB |
```